Add a edge

Update dataframe for visNetwork graph

Usage

add_edge(id = NULL, from, to, label, vals, create_link = TRUE)

Arguments

id

Edge id. If create_link is FALSE, id cannot be not NULL.

from

Node id.

to

Node it.

label

Edge label. This is useful to map the existing connected nodes to the input slots of the receiving node (for instance a join block).

vals

Global vals reactive values. Read-write access.

create_link

Create a link in the board?