Dynamically group nodes

Given a set of selected nodes, add them to a unique group and apply unique color and labels.

Useful when removing a stack.

Usage

stack_nodes(stack_id = NULL, nodes = NULL, rv, parent, session)

unstack_nodes(parent, session)

Arguments

stack_id

Stack id to attach nodes to.

nodes

Vector of node ids to stack.

rv

Board reactive values.

parent

Global scope (entire app) reactive values.

session

Shiny session object.